# server/docs/swagger.yaml
definitions:
  community.SetAvatarResponse:
    properties:
      url:
        type: string
    type: object
  domain.Community:
    properties:
      avatar:
        type: string
      config:
        $ref: '#/definitions/domain.CommunityConfig'
      description:
        type: string
      id:
        type: string
      isMember:
        type: boolean
      members:
        items:
          $ref: '#/definitions/domain.Member'
        type: array
      membersCount:
        type: integer
      name:
        type: string
      tgChatID:
        type: integer
    type: object
  domain.CommunityConfig:
    properties:
      fields:
        items:
          $ref: '#/definitions/domain.Field'
        type: array
    type: object
  domain.CreateCommunity:
    properties:
      avatar:
        type: string
      config:
        $ref: '#/definitions/domain.CommunityConfig'
      description:
        type: string
      name:
        type: string
    type: object
  domain.Field:
    properties:
      description:
        type: string
      textarea:
        $ref: '#/definitions/domain.FieldTextarea'
      textinput:
        $ref: '#/definitions/domain.FieldTextinput'
      title:
        type: string
      type:
        $ref: '#/definitions/domain.FieldType'
    type: object
  domain.FieldTextarea:
    properties:
      default:
        type: string
    type: object
  domain.FieldTextinput:
    properties:
      default:
        type: string
    type: object
  domain.FieldType:
    enum:
    - textinput
    - textarea
    type: string
    x-enum-varnames:
    - FieldTypeTextinput
    - FieldTypeTextarea
  domain.FieldValue:
    properties:
      textarea:
        $ref: '#/definitions/domain.FieldValueTextarea'
      textinput:
        $ref: '#/definitions/domain.FieldValueTextinput'
      type:
        $ref: '#/definitions/domain.FieldType'
    type: object
  domain.FieldValueTextarea:
    properties:
      value:
        type: string
    type: object
  domain.FieldValueTextinput:
    properties:
      value:
        type: string
    type: object
  domain.Member:
    properties:
      community:
        $ref: '#/definitions/domain.Community'
      config:
        $ref: '#/definitions/domain.MemberConfig'
      isAdmin:
        type: boolean
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.MemberConfig:
    properties:
      fields:
        additionalProperties:
          $ref: '#/definitions/domain.FieldValue'
        description: title to field
        type: object
    type: object
  domain.PatchCommunity:
    properties:
      avatar:
        type: string
      config:
        $ref: '#/definitions/domain.CommunityConfig'
      description:
        type: string
      id:
        type: string
      name:
        type: string
      tgChatID:
        type: integer
    type: object
  domain.PatchMe:
    properties:
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
    type: object
  domain.PatchMember:
    properties:
      config:
        $ref: '#/definitions/domain.MemberConfig'
      id:
        type: string
      isAdmin:
        type: boolean
      userId:
        type: string
    type: object
  domain.User:
    properties:
      avatar:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      telegramId:
        type: integer
      telegramUsername:
        type: string
    type: object
  domain.UserProfile:
    properties:
      members:
        items:
          $ref: '#/definitions/domain.Member'
        type: array
      user:
        $ref: '#/definitions/domain.User'
    type: object
info:
  contact: {}
  description: Manage chats, users
  title: TGLinked server
  version: "1.0"
paths:
  /communities:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Communities data
          schema:
            items:
              $ref: '#/definitions/domain.Community'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get community previews
      tags:
      - communities
    post:
      consumes:
      - application/json
      parameters:
      - description: Community data
        in: body
        name: community
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCommunity'
      produces:
      - application/json
      responses:
        "200":
          description: Community data
          schema:
            $ref: '#/definitions/domain.Community'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create community
      tags:
      - communities
  /communities/id/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Community data
          schema:
            $ref: '#/definitions/domain.Community'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get community
      tags:
      - communities
    patch:
      consumes:
      - application/json
      parameters:
      - description: Community data
        in: body
        name: community
        required: true
        schema:
          $ref: '#/definitions/domain.PatchCommunity'
      produces:
      - application/json
      responses:
        "200":
          description: Community data
          schema:
            $ref: '#/definitions/domain.Community'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Patch community
      tags:
      - communities
  /communities/id/{id}/join:
    post:
      consumes:
      - application/json
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      - description: Community config
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/domain.MemberConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Join to community
      tags:
      - communities
  /communities/id/{id}/leave:
    post:
      consumes:
      - application/json
      parameters:
      - description: community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Leave community
      tags:
      - communities
  /communities/id/{id}/members/id/{member_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      - description: Member id
        in: path
        name: member_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Member
          schema:
            $ref: '#/definitions/domain.Member'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get member
      tags:
      - communities
    patch:
      consumes:
      - application/json
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: member_id
        required: true
        type: string
      - description: Community config
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/domain.PatchMember'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Patch member
      tags:
      - communities
  /communities/id/{id}/members/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Members
          schema:
            items:
              $ref: '#/definitions/domain.Member'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Search members
      tags:
      - communities
  /communities/id/{id}/preview:
    get:
      consumes:
      - application/json
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Community preview
          schema:
            $ref: '#/definitions/domain.Community'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get community preview
      tags:
      - communities
  /communities/id/{id}/set_avatar:
    post:
      consumes:
      - application/json
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      - description: Image data
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: A url to the stored image
          schema:
            $ref: '#/definitions/community.SetAvatarResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Set avatar for community
      tags:
      - communities
  /communities/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Communities
          schema:
            items:
              $ref: '#/definitions/domain.Community'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Search communities
      tags:
      - communities
  /users/me:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Delete me
      tags:
      - users
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/domain.UserProfile'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Get me
      tags:
      - users
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create me
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.PatchMe'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Update me
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Token
    type: apiKey
swagger: "2.0"